Implemented requested changes including renaming the lib packages to internal.
claell)
