+++ addons/binding/org.openhab.binding.yeelight/about.html
@@ -23,17 +23,5 @@
 		.
 	</p>
 
-
-	<h3>Third Party Content</h3>
-	<p>The Content includes items that have been sourced from third parties as set out below. If you did not receive this Content directly from the Eclipse Foundation, the following is provided for informational purposes only, and you should look to the Redistributor's license for terms and conditions of use.</p>
-	<p>
-		<strong>yeelight_lib Library</strong> <br />
-		<br /> This bundle includes yeelight library for sending control command to Yeelight devices. The sources can be found at
-		<a href="https://github.com/coasterli/YeelightLib">GitHub</a>
-		and are made available under the
-		<a href="about_files/LICENSE">MIT license</a>
-		. However the version linked above seems to be outdated. The sources are included in lib/yeelight_lib_original_with_sources.jar.
-	</p>
-
 </body>
 </html>
