+++ addons/binding/org.openhab.binding.yeelight/build.properties
@@ -5,5 +5,3 @@ bin.includes = META-INF/,\
                OSGI-INF/,\
                ESH-INF/,\
                about.html,\
-               about_files/
-
