+++ addons/binding/org.openhab.binding.yeelight/src/main/java/org/openhab/binding/yeelight/handler/YeelightHandlerBase.java
@@ -26,16 +26,16 @@ import org.eclipse.smarthome.core.thing.binding.BaseThingHandler;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.openhab.binding.yeelight.YeelightBindingConstants;
-import org.openhab.binding.yeelight.lib.device.ConnectState;
-import org.openhab.binding.yeelight.lib.device.DeviceBase;
-import org.openhab.binding.yeelight.lib.device.DeviceFactory;
-import org.openhab.binding.yeelight.lib.device.DeviceStatus;
-import org.openhab.binding.yeelight.lib.enums.DeviceAction;
-import org.openhab.binding.yeelight.lib.enums.DeviceMode;
-import org.openhab.binding.yeelight.lib.enums.DeviceType;
-import org.openhab.binding.yeelight.lib.listeners.DeviceConnectionStateListener;
-import org.openhab.binding.yeelight.lib.listeners.DeviceStatusChangeListener;
-import org.openhab.binding.yeelight.lib.services.DeviceManager;
+import org.openhab.binding.yeelight.internal.lib.device.ConnectState;
+import org.openhab.binding.yeelight.internal.lib.device.DeviceBase;
+import org.openhab.binding.yeelight.internal.lib.device.DeviceFactory;
+import org.openhab.binding.yeelight.internal.lib.device.DeviceStatus;
+import org.openhab.binding.yeelight.internal.lib.enums.DeviceAction;
+import org.openhab.binding.yeelight.internal.lib.enums.DeviceMode;
+import org.openhab.binding.yeelight.internal.lib.enums.DeviceType;
+import org.openhab.binding.yeelight.internal.lib.listeners.DeviceConnectionStateListener;
+import org.openhab.binding.yeelight.internal.lib.listeners.DeviceStatusChangeListener;
+import org.openhab.binding.yeelight.internal.lib.services.DeviceManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
