+++ addons/binding/org.openhab.binding.yeelight/src/main/java/org/openhab/binding/yeelight/handler/YeelightStripeHandler.java
@@ -11,8 +11,7 @@ package org.openhab.binding.yeelight.handler;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
-
-import org.openhab.binding.yeelight.lib.device.DeviceStatus;
+import org.openhab.binding.yeelight.internal.lib.device.DeviceStatus;
 
 /**
  * The {@link YeelightStripeHandler} is responsible for handling commands, which are
