+++ addons/binding/org.openhab.binding.yeelight/src/main/java/org/openhab/binding/yeelight/handler/YeelightWhiteHandler.java
@@ -13,10 +13,9 @@ import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
 import org.openhab.binding.yeelight.YeelightBindingConstants;
+import org.openhab.binding.yeelight.internal.lib.device.DeviceStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import org.openhab.binding.yeelight.lib.device.DeviceStatus;
 
 /**
  * The {@link YeelightWhiteHandler} is responsible for handling commands, which are
