+++ addons/binding/org.openhab.binding.yeelight/src/main/java/org/openhab/binding/yeelight/internal/lib/device/ConnectState.java
@@ -6,13 +6,19 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.yeelight.lib.listeners;
-
-import org.openhab.binding.yeelight.lib.device.DeviceStatus;
+package org.openhab.binding.yeelight.internal.lib.device;
 
 /**
+ * The {@link ConnectState} lists different connection states.
+ *
  * @author Coaster Li - Initial contribution
  */
-public interface DeviceStatusChangeListener {
-    void onStatusChanged(String prop, DeviceStatus status);
+public enum ConnectState {
+
+    DISCONNECTED,
+
+    CONNECTTING,
+
+    CONNECTED;
+
 }
