+++ addons/binding/org.openhab.binding.yeelight/src/main/java/org/openhab/binding/yeelight/internal/lib/enums/MethodAction.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.yeelight.lib.enums;
+package org.openhab.binding.yeelight.internal.lib.enums;
 
 /**
  * @author Coaster Li - Initial contribution
