--- addons/binding/org.openhab.binding.yeelight/src/main/java/org/openhab/binding/yeelight/lib/enums/DeviceMode.java
@@ -1,40 +1,0 @@
-/**
- * Copyright (c) 2010-2018 by the respective copyright holders.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.openhab.binding.yeelight.lib.enums;
-
-/**
- * @author Coaster Li - Initial contribution
- */
-public enum DeviceMode {
-
-    /*
-     * color temperature mode
-     */
-    MODE_SUNHINE,
-
-    /*
-     * color mode
-     */
-    MODE_COLOR,
-
-    /*
-     * color mode
-     */
-    MODE_HSV,
-
-    /*
-     * color flow mode
-     */
-    MODE_COLORFLOW,
-
-    /*
-     * device is off.
-     */
-    MODE_OFF;
-}
